Updated to fix docker and passing the SCD tests
