--- Dockerfile-dev
@@ -1,11 +1,0 @@
-
-FROM --platform=linux/amd64 python:3.12-slim
-ENV PYTHONUNBUFFERED=1
-
-
-RUN mkdir /app
-WORKDIR /app
-COPY requirements.txt /app/
-RUN pip install -r requirements.txt
-RUN rm -rf __pycache__
-COPY . /app/
