+++ constraint_operations/dss_constraints_helper.py
@@ -7,6 +7,7 @@ from os import environ as env
 import requests
 import urllib3
 from dacite import from_dict
+from dotenv import find_dotenv, load_dotenv
 
 from auth_helper import dss_auth_helper
 from common.auth_token_audience_helper import generate_audience_from_base_url
@@ -14,8 +15,6 @@ from common.database_operations import (
     FlightBlenderDatabaseReader,
     FlightBlenderDatabaseWriter,
 )
-from geo_fence_operations.data_definitions import GeofencePayload
-from scd_operations.dss_scd_helper import VolumesConverter
 from scd_operations.scd_data_definitions import Time, Volume4D
 
 from .data_definitions import (
@@ -32,65 +31,6 @@ if ENV_FILE:
     load_dotenv(ENV_FILE)
 
 logger = logging.getLogger("django")
-
-
-class ConstraintsWriter:
-    def __init__(self) -> None:
-        self.my_database_reader = FlightBlenderDatabaseReader()
-        self.my_database_writer = FlightBlenderDatabaseWriter()
-
-    # def parse_stored_constraint_details(self, geozone_id: str) -> ConstraintDetails | None:
-    #     pass
-
-    # def parse_and_load_stored_constraint_reference(self, geozone_id: str) -> ConstraintReference | None:
-    #     pass
-
-    def write_nearby_constraints(self, constraints: list[Constraint]):
-        # This method writes the constraint reference and constraint details to the database
-        my_volumes_converter = VolumesConverter()
-        for constraint in constraints:
-            constraint_reference = constraint.reference
-            constraint_details = constraint.details
-            # Check if the constraint reference already exists in the database
-            constraint_reference_exists = self.my_database_reader.check_constraint_reference_id_exists(
-                constraint_reference_id=str(constraint_reference.id)
-            )
-            geofence_id = str(uuid.uuid4())
-
-            if constraint_reference_exists:
-                geo_fence = self.my_database_reader.get_geofence_by_constraint_reference_id(constraint_reference_id=str(constraint_reference.id))
-                geofence_id = geo_fence.id
-
-            my_volumes_converter.convert_volumes_to_geojson(volumes=constraint_details.volumes)
-
-            geofence_payload = GeofencePayload(
-                id=geofence_id,
-                raw_geo_fence=my_volumes_converter.geojson,
-                upper_limit=my_volumes_converter.upper_altitude,
-                lower_limit=my_volumes_converter.upper_altitude,
-                altitude_ref=my_volumes_converter.altitude_ref,
-                name=constraint_details.volumes[0].name,
-                bounds=my_volumes_converter.get_bounds(),
-                status=1,
-                message="Constraint from peer USS",
-                is_test_dataset=False,
-                start_datetime=constraint_reference.time_start,
-                end_datetime=constraint_reference.time_end,
-            )
-
-            geo_fence = self.my_database_writer.create_or_update_geofence(geofence_payload=geofence_payload)
-            # Create a new ConstraintReference object
-
-            self.my_database_writer.create_or_update_constraint_reference(
-                constraint_reference=constraint_reference,
-                geofence=geo_fence,
-            )
-
-            # Write the constraint details to the database
-            self.my_database_writer.create_or_update_constraint_detail(
-                constraint=constraint_details,
-                geofence=geo_fence,
-            )
 
 
 class ConstraintOperations:
