+++ docker-compose.yml
@@ -3,7 +3,7 @@ services:
   redis-blender:
     platform: linux/amd64
     command: ["redis-server", "/redis.conf", "--requirepass", "$REDIS_PASSWORD"]
-    image: "redis:latest"
+    image: "valkey/valkey:latest"
     expose:
       - "6379"
     volumes:
@@ -43,8 +43,8 @@ services:
     depends_on:
       - redis-blender
       - db-blender
-    # volumes:
-    #   - .:/app
+    volumes:
+      - .:/app
     networks:
       - interop_ecosystem_network
